+++ src/main/scala/com/rayrobdod/json/builder/Builder.scala
@@ -125,4 +125,42 @@ trait Builder[-Key, -Value, +Failure, Subject] {
 		}
 	}
 	
+	/**
+	 * Create a Builder which builds two items from the same input. The first
+	 * item built being the one that `this` would build, and the second item
+	 * built being the one that `that` would build.
+	 * 
+	 * Most useful if the Parser input is neither stable nor resendable - say a network stream.
+	 * 
+	 * @param that the other builder
+	 * @since 3.1
+	 */
+	final def zip[K2, V2, F2, S2](that:Builder[K2, V2, F2, S2])(implicit evk: K2 <:< Key, evv: V2 <:< Value, evf: Failure <:< F2):Builder[K2,V2,F2,(Subject, S2)] = new Builder[K2,V2,F2,(Subject, S2)] {
+		private[this] val astBuilder = MapBuilder.apply[K2, V2]
+		private[this] val ast2Parser = new com.rayrobdod.json.parser.RecursiveMapParser[K2, V2]
+		private[this] val ident2Parser = new com.rayrobdod.json.parser.IdentityParser[V2]
+		private[this] val ast1Parser = ast2Parser.mapKey(evk).mapValue(evv)
+		private[this] val ident1Parser = ident2Parser.mapValue(evv)
+		
+		override def init:(Subject, S2) = ((Builder.this.init, that.init))
+		override def apply[Input, PF](a:(Subject, S2), key:K2, b:Input, c:Parser[K2, V2, PF, Input]):ParserRetVal[(Subject, S2), Nothing, PF, F2] = {
+			// `b` might be mutable, so `c.parse(…, c)` can only be called once;
+			// `ast` or `value` however are not, so those may be parsed multiple times 
+			c.parse(astBuilder, b).fold(
+				{ast:Map[K2, Either[MapBuilder.RecursiveSubject[K2, V2], V2]] =>
+					val e1:ParserRetVal[Subject, Nothing, PF, F2] = Builder.this.apply(a._1, key, ast, ast1Parser).builderFailure.map(evf)
+					val e2:ParserRetVal[S2, Nothing, PF, F2] = that.apply(a._2, key, ast, ast2Parser)
+					for {r1 <- e1.complex; r2 <- e2.complex} yield {((r1, r2))}
+				},
+				{value:V2 => 
+					val e1 = Builder.this.apply(a._1, key, value, ident1Parser).builderFailure.map(evf)
+					val e2 = that.apply(a._2, key, value, ident2Parser)
+					for {r1 <- e1.complex; r2 <- e2.complex} yield {((r1, r2))}
+				},
+				ParserRetVal.ParserFailure.apply _,
+				ParserRetVal.BuilderFailure.apply _
+			)
+		}
+	}
+	
 }