+++ src/main/scala/com/rayrobdod/json/builder/PrettyJsonBuilder.scala
@@ -128,6 +128,28 @@ object PrettyJsonBuilder {
 	
 	
 	/**
+	 * Shorthand for a PrettyJsonBuilder using a MinifiedPrettyParams
+	 * @since 3.1
+	 */
+	def minified(charset:Charset = UTF_8):Builder[StringOrInt, JsonValue, PrettyJsonBuilder.Failures, String] = new PrettyJsonBuilder(MinifiedPrettyParams, charset)
+	/**
+	 * Shorthand for a PrettyJsonBuilder using an IndentPrettyParams using two spaces for the indent
+	 * @since 3.1
+	 */
+	def space2(charset:Charset = UTF_8):Builder[StringOrInt, JsonValue, PrettyJsonBuilder.Failures, String] = new PrettyJsonBuilder(new IndentPrettyParams("  "), charset)
+	/**
+	 * Shorthand for a PrettyJsonBuilder using an IndentPrettyParams using four spaces for the indent
+	 * @since 3.1
+	 */
+	def space4(charset:Charset = UTF_8):Builder[StringOrInt, JsonValue, PrettyJsonBuilder.Failures, String] = new PrettyJsonBuilder(new IndentPrettyParams("    "), charset)
+	/**
+	 * Shorthand for a PrettyJsonBuilder using an IndentPrettyParams using a tab for the indent
+	 * @since 3.1
+	 */
+	def tabbed(charset:Charset = UTF_8):Builder[StringOrInt, JsonValue, PrettyJsonBuilder.Failures, String] = new PrettyJsonBuilder(new IndentPrettyParams(), charset)
+	
+	
+	/**
 	 * Possible failures that can occur in a PrettyJsonBuilder
 	 * @since 4.0
 	 */
@@ -146,7 +168,6 @@ object PrettyJsonBuilder {
 		final case class ArrayKeyNotIncrementing(recieved:Int, expecting:Int) extends Failures
 		object IllegalFoldingInBuilder extends Failures
 	}
-	
 	
 	
 	/**