+++ src/main/scala/com/rayrobdod/json/parser/CborParser.scala
@@ -339,7 +339,7 @@ object CborParser {
 	
 	/**
 	 * A function that is parameterized at the function level instead of the class level
-	 * @since next
+	 * @since 3.1
 	 */
 	trait TagFunction {
 		def apply[A, BF](b:Builder[CborValue, CborValue, BF, A], i:DataInput):ParseReturnValue[A, BF] 
@@ -347,7 +347,7 @@ object CborParser {
 	
 	/**
 	 * A partial function that takes a Cbor tag number and returns a function that builds the value described by the tags
-	 * @since next
+	 * @since 3.1
 	 */
 	trait TagMatcher {
 		def unapply(tag:Long):Option[TagFunction]
@@ -364,7 +364,7 @@ object CborParser {
 	
 	/**
 	 * Built-in TagMatchers
-	 * @since next
+	 * @since 3.1
 	 */
 	object TagMatcher {
 		
@@ -502,6 +502,10 @@ object CborParser {
 		final val END_OF_LIST:Byte = 31
 	}
 	
+	/**
+	 * Known tag codes.
+	 * Because magic numbers are bad.
+	 */
 	private[json] object TagCodes {
 		final val POS_BIG_INT:Byte = 2
 		final val NEG_BIG_INT:Byte = 3